pyproject.toml
[project]
name = "gitverse-mirror"
version = "0.1.0"
description = "Git repository mirroring tool"
requires-python = ">=3.11"
dependencies = [
    "schedule>=1.2",
    "pyyaml>=6.0",
    "coloredlogs>=15.0",
]

[project.scripts]
gitverse-mirror = "reflection.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/reflection"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "mypy>=1.10",
]
